+++ Modules/Intent.Modules.Common.CSharp/FactoryExtensions/CSharpFileBuilderFactoryExtension.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using System.Threading;
@@ -21,59 +22,89 @@ namespace Intent.Modules.Common.Plugins
 
         protected override void OnAfterTemplateRegistrations(IApplication application)
         {
-            var templates = application.OutputTargets.SelectMany(x => x.TemplateInstances)
-                .Where(x => x.CanRunTemplate())
-                .OfType<ICSharpFileBuilderTemplate>()
-                .ToList();
+            PerformConfiguration(
+                context: application,
+                getActions: t => t.CSharpFile.GetConfigurationDelegates(),
+                afterAllComplete: null);
 
-            templates.ForEach(x =>
-            {
-                try
-                {
-                    x.CSharpFile.StartBuild();
-                }
-                catch (ElementException)
-                {
-                    throw;
-                }
-                catch (Exception e)
-                {
-                    if (x.TryGetModel<IElementWrapper>(out var model))
-                    {
-                        throw new ElementException(model.InternalElement, $@"An unexpected error occurred while building C# file [{x.Namespace}.{x.ClassName}] from template [{x.Id}]. See inner exception for more details:", e);
-                    }
-                    throw new Exception($@"An unexpected error occurred while building C# file [{x.Namespace}.{x.ClassName}] from template [{x.Id}]. See inner exception for more details:", e);
-                }
-            });
-            templates.ForEach(x =>
-            {
-                try
-                {
-                    x.CSharpFile.CompleteBuild();
-                }
-                catch (ElementException)
-                {
-                    throw;
-                }
-                catch (Exception e)
-                {
-                    if (x.TryGetModel<IElementWrapper>(out var model))
-                    {
-                        throw new ElementException(model.InternalElement, $@"An unexpected error occurred while building C# file [{x.Namespace}.{x.ClassName}] from template [{x.Id}]. See inner exception for more details:", e);
-                    }
-                    throw new Exception($@"An unexpected error occurred while building C# file [{x.Namespace}.{x.ClassName}] from template [{x.Id}]. See inner exception for more details:", e);
-                }
-            });
+            PerformConfiguration(
+                context: application,
+                getActions: t => t.CSharpFile.GetConfigurationDelegates(),
+                afterAllComplete: t => t.CSharpFile.MarkCompleteBuildAsDone());
         }
 
         protected override void OnBeforeTemplateExecution(IApplication application)
         {
-            var templates = application.OutputTargets.SelectMany(x => x.TemplateInstances)
+            PerformConfiguration(
+                context: application,
+                getActions: t => t.CSharpFile.GetConfigurationAfterDelegates(),
+                afterAllComplete: t => t.CSharpFile.MarkAfterBuildAsDone());
+        }
+
+        private static void PerformConfiguration(
+            ISoftwareFactoryExecutionContext context,
+            Func<ICSharpFileBuilderTemplate, IReadOnlyCollection<(Action Invoke, int Order)>> getActions,
+            Action<ICSharpFileBuilderTemplate> afterAllComplete)
+        {
+            var cSharpFileBuilderTemplates = context.OutputTargets.SelectMany(x => x.TemplateInstances)
                 .Where(x => x.CanRunTemplate())
                 .OfType<ICSharpFileBuilderTemplate>()
-                .ToList();
+                .ToArray();
 
-            templates.ForEach(x => x.CSharpFile.AfterBuild());
+            while (true)
+            {
+                var actions = cSharpFileBuilderTemplates
+                    .SelectMany(
+                        collectionSelector: getActions,
+                        resultSelector: (template, configuration) => new
+                        {
+                            Template = template,
+                            TemplateType = template.GetType().ToString(),
+                            configuration.Order,
+                            configuration.Invoke
+                        })
+                    .OrderBy(x => x.Order)
+                    .ThenBy(x => x.TemplateType)
+                    .ToArray();
+
+                if (actions.Length == 0)
+                {
+                    break;
+                }
+
+                foreach (var action in actions)
+                {
+                    try
+                    {
+                        action.Invoke();
+                    }
+                    catch (ElementException)
+                    {
+                        throw;
+                    }
+                    catch (Exception e)
+                    {
+                        var template = action.Template;
+
+                        if (template.TryGetModel<IElementWrapper>(out var model))
+                        {
+                            throw new ElementException(model.InternalElement, $"An unexpected error occurred while building C# file [{template.Namespace}.{template.ClassName}] from template [{template.Id}]. See inner exception for more details:", e);
+                        }
+
+                        throw new Exception($"An unexpected error occurred while building C# file [{template.Namespace}.{template.ClassName}] from template [{template.Id}]. See inner exception for more details:", e);
+                    }
+                }
+            }
+
+            if (afterAllComplete == null)
+            {
+                return;
+            }
+
+            foreach (var template in cSharpFileBuilderTemplates)
+            {
+                afterAllComplete(template);
+            }
         }
     }
 }